build clean command added.

+++ wasanbon/core/rtc/build_rtc.py
@@ -25,8 +25,20 @@ def clean_rtc_win32(rtcp):
     current_y = yaml.load(open(os.path.join(os.getcwd(), 'setting.yaml'), 'r'))
     rtc_dir, rtc_xml = os.path.split(rtcp.getRTCProfileFileName())
     build_dir = os.path.join(rtc_dir, 'build-win32')
-    print 'removing %s' % build_dir
-    shutil.rmtree(build_dir, ignore_errors=True )
+    if build_dir.startswith(os.getcwd()):
+        build_dir = build_dir[len(os.getcwd())+1:]
+    if os.path.isdir(build_dir):
+        print 'removing %s' % build_dir
+        shutil.rmtree(build_dir, ignore_errors=True )
+
+    for root, dirs, files in os.walk(rtc_dir):
+        for file in files:
+            if file.endswith('~'):
+                fullpath = os.path.join(root, file)
+                if fullpath.startswith(os.getcwd()):
+                    fullpath = fullpath[len(os.getcwd())+1]
+                print 'removing %s' % fullpath
+                os.remove(os.path.join(root, file))
 
 def build_rtc_win32(rtcp):
     setting = load_settings()
